remove unnecessary/annoying prefix= arg
in most contexts this library might be used (e.g. CGI/WSGI environ)
it is quite annoying to have to suppress the default prefix,
and it hurts the API's simplicity. I used to worry that this would
help perpetuate confusion about the difference between request and
response formats but this isn't the right way to fix that.

since this breaks backward compatibility with scripts that specify
prefix=, I am incrementing the major version.