add new ways to change Cookies' Cookie class
This was spurred by a suggestion by Laurent Bachelier (thanks, Laurent).

For those just tuning in, Cookies is an class which offers dict-like access to
collections of Cookie objects, each representing a name/value pair with its own
attributes (corresponding to attributes of an HTTP Cookie).

To make a cookie collection which constructs instances using your own class
(suppose it is called Scone) - use Cookies(_cookie_class=Scone).

To make a Cookies subclass generating instances which do the same thing,
override DEFAULT_COOKIE_CLASS on the subclass itself, and instantiate it
as normal.

It's also still possible to use the lower-level API to make whatever collection
interface you really want to use, if your requirements are really different.