add 3.4 and drop 3.1 (tox/pip breakage)
A backward-incompatible version bump comes with this.

None of the supporting test tools seem to actually support it any more
and those regressions show no signs of being fixed. 3.1 is a royal pain
anyway, not important in any distro I know of, and it was going to happen
someday, so why not now.
3.1 may still work, I'm just not interested in implied guarantees when
testing it is such a pain.

Sequenced this after _cookie_class so that can be in the new version