stop parse_domain from stripping initial dot
Hmm.

Unlike spaces and quotes, this initial dot character could carry information
that might be of interest to the caller who is parsing a Set-Response line.

Arguably it should be up to the user-agent itself, not the parsing library
called by the user-agent, to ignore the initial dot.

After all, I went far enough to leave the initial dot in DOMAIN_AV regexp specifically to
let this initial period through... and also wrote in a comment:

    # Domains starting with periods are not RFC-valid, but this is very common
    # in existing cookies, so they should still parse with DOMAIN_AV.

Then I wrote DOMAIN_AV so that the attribute would parse without vomiting.

But perhaps, as a separate change, generated renders should strip the dot.
At least via the default renderer.