declare initial dot domains valid by fiat
This really, really sucks. But I am very aggressively not allowing
attribute values I define as invalid to exist on Cookie objects,
which prevents even temporarily checking if there is an initial dot.
Domain names with initial dot are NOT RFC valid and we shouldn't
produce them, but we might want to see them so they are "valid"
as far as that goes.